requirements.txt
tweepy==4.14.0
transformers==4.40.0
torch>=1.12.0
pymongo==4.4.0
python-dotenv==1.0.0
streamlit==1.25.0
plotly==5.15.0
fastapi==0.95.0
uvicorn==0.22.0
pandas==2.2.0
reportlab==4.1.0
python-dateutil==2.8.2
scikit-learn
matplotlib